Update c program to find factorial.cpp

+++ C_Programs/Lab3/c program to find factorial.cpp
@@ -2,23 +2,18 @@
 
 int main() {
     int n;
-    unsigned long long factorial = 1;
-
-    // Input the number from the user
+    long factorial = 1;
     printf("Enter a positive integer: ");
     scanf("%d", &n);
-
-    // Check if the input is a negative number
     if (n < 0) {
         printf("Factorial is not defined for negative numbers.\n");
-    } else {
-        // Calculate the factorial
+    }
+    else {
+
         for (int i = 1; i <= n; i++) {
             factorial *= i;
         }
-
-        // Print the factorial
-        printf("The factorial of %d is %llu\n", n, factorial);
+        printf("The factorial of %d is %li\n", n, factorial);
     }
 
     return 0;
